% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jaccard_similarity_heatmap.R
\name{plot_jaccard_similarity_heatmap}
\alias{plot_jaccard_similarity_heatmap}
\title{Plots a heatmap of the Jaccard similarity between clusters created with two different methods.}
\usage{
plot_jaccard_similarity_heatmap(clusters1, clusters2, width = 10, height = 10)
}
\arguments{
\item{clusters1}{A vector named by sequence IDs with values being subtrees defining the cluster.}

\item{clusters2}{A vector named by the same sequence IDs as clusters1 with values being subtrees
defining the cluster.}

\item{width}{The width of the heatmap plot.}

\item{height}{The height of the heatmap plot.}
}
\value{
A ggplot plot object showing the overlap between clusters.
}
\description{
Plots a heatmap of the Jaccard similarity between clusters created with two different methods.
}
